backend/requirements.txt
# FastAPI backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlmodel>=0.0.14

# Validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# LLM integration (Gemini via Replit AI Integrations)
google-genai>=1.0.0
tenacity>=8.0.0

# PDF processing
pdfplumber>=0.10.0
python-multipart>=0.0.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0